Add VirtualButton to replace buttons in banks
The current push button implementations for 1,2,3 and 6 (ladder) buttons
are very different from one another, leading to inconsistent and hard to
understand usage.

This commit introduces the concept of "virtual" buttons as a new
layer. Pedals only define actual buttons (1 to 6), and what the buttons
do is completely defined by the new layer.

For the moment, functionality is limited to sending direct MIDI messages
and up/down, but all others should be fairly easy to add. After that, it
can replace the current "banks" page.